chan: channel 独立为一个子模块
